federated-learning/requirements.txt
# Core dependencies
grpcio==1.57.0
grpcio-tools==1.57.0
numpy==1.24.3
tensorflow==2.13.0
protobuf==4.21.6

# Web server
flask==2.3.3
flask-cors==4.0.0

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0

# Database
sqlite3==2.6.0

# NEAR interaction
near-api-py==0.1.0
aiohttp==3.8.5
base58==2.1.1

# Async support
asyncio==3.4.3

# Testing and development
pytest==7.4.0
pytest-asyncio==0.21.1 